feat: add new `InputInvoiceMessageContent` type
docs: describe all kind of panics for `tip::Tip::suggested_tips`
