+++ src/bot.rs
@@ -25,10 +25,10 @@ use crate::{
         keyboard::inline,
         message,
         parameters::{
-            poll, AllowedUpdates, CallbackAction, ImplicitChatId, Text,
+            poll, AllowedUpdates, CallbackAction, ImplicitChatId, Invoice, Text,
         },
         passport, pre_checkout_query, shipping, user, BotCommand,
-        InlineMessageId, LabeledPrice,
+        InlineMessageId,
     },
 };
 use std::{net::IpAddr, num::NonZeroU32, sync::Arc};
@@ -626,23 +626,9 @@ impl Bot {
     pub fn send_invoice(
         &self,
         chat_id: impl Into<chat::Id>,
-        title: impl Into<String>,
-        description: impl Into<String>,
-        payload: impl Into<String>,
-        provider_token: impl Into<String>,
-        currency: impl Into<String>,
-        prices: impl Into<Vec<LabeledPrice>>,
+        invoice: Invoice,
     ) -> SendInvoice<'_> {
-        SendInvoice::new(
-            &self.inner,
-            chat_id,
-            title,
-            description,
-            payload,
-            provider_token,
-            currency,
-            prices,
-        )
+        SendInvoice::new(&self.inner, chat_id, invoice)
     }
 
     /// Sends a location.
