+++ src/contexts/methods/message.rs
@@ -11,8 +11,8 @@ use crate::{
         },
         keyboard::inline,
         message,
-        parameters::{poll, ImplicitChatId, Text},
-        user, LabeledPrice,
+        parameters::{poll, ImplicitChatId, Invoice, Text},
+        user,
     },
 };
 
@@ -296,46 +296,14 @@ pub trait Message: fields::Message {
 
     /// Sends an invoice to this chat.
     #[allow(clippy::too_many_arguments)]
-    fn send_invoice(
-        &self,
-        title: impl Into<String>,
-        description: impl Into<String>,
-        payload: impl Into<String>,
-        provider_token: impl Into<String>,
-        currency: impl Into<String>,
-        prices: impl Into<Vec<LabeledPrice>>,
-    ) -> SendInvoice<'_> {
-        self.bot().send_invoice(
-            self.chat().id,
-            title,
-            description,
-            payload,
-            provider_token,
-            currency,
-            prices,
-        )
+    fn send_invoice(&self, invoice: Invoice) -> SendInvoice<'_> {
+        self.bot().send_invoice(self.chat().id, invoice)
     }
 
     /// Sends an invoice in reply to this message.
     #[allow(clippy::too_many_arguments)]
-    fn send_invoice_in_reply(
-        &self,
-        title: impl Into<String>,
-        description: impl Into<String>,
-        payload: impl Into<String>,
-        provider_token: impl Into<String>,
-        currency: impl Into<String>,
-        prices: impl Into<Vec<LabeledPrice>>,
-    ) -> SendInvoice<'_> {
-        self.send_invoice(
-            title,
-            description,
-            payload,
-            provider_token,
-            currency,
-            prices,
-        )
-        .in_reply_to(self.message_id())
+    fn send_invoice_in_reply(&self, invoice: Invoice) -> SendInvoice<'_> {
+        self.send_invoice(invoice).in_reply_to(self.message_id())
     }
 
     /// Sends a location to this chat.
