+++ src/methods/send_invoice.rs
@@ -6,8 +6,7 @@ use crate::{
         chat,
         keyboard::inline,
         message::{self, Message},
-        parameters::{tip, Photo},
-        LabeledPrice,
+        parameters::Invoice,
     },
 };
 use serde::Serialize;
@@ -23,32 +22,8 @@ pub struct SendInvoice<'a> {
     #[serde(skip)]
     bot: &'a InnerBot,
     chat_id: chat::Id,
-    title: String,
-    description: String,
-    payload: String,
-    provider_token: String,
-    currency: String,
-    prices: Vec<LabeledPrice>,
     #[serde(skip_serializing_if = "Option::is_none")]
     start_parameter: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    provider_data: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none", flatten)]
-    photo: Option<Photo>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    need_name: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    need_phone_number: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    need_email: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    need_shipping_address: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    send_phone_number_to_provider: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    send_email_to_provider: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    is_flexible: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
     disable_notification: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -57,7 +32,7 @@ pub struct SendInvoice<'a> {
     #[serde(skip_serializing_if = "Option::is_none")]
     reply_markup: Option<inline::Keyboard>,
     #[serde(flatten)]
-    tip: Option<tip::Tip>,
+    invoice: Invoice,
 }
 
 impl<'a> SendInvoice<'a> {
@@ -65,108 +40,18 @@ impl<'a> SendInvoice<'a> {
     pub(crate) fn new(
         bot: &'a InnerBot,
         chat_id: impl Into<chat::Id>,
-        title: impl Into<String>,
-        description: impl Into<String>,
-        payload: impl Into<String>,
-        provider_token: impl Into<String>,
-        currency: impl Into<String>,
-        prices: impl Into<Vec<LabeledPrice>>,
+        invoice: Invoice,
     ) -> Self {
         Self {
             bot,
             chat_id: chat_id.into(),
-            title: title.into(),
-            description: description.into(),
-            payload: payload.into(),
-            provider_token: provider_token.into(),
-            currency: currency.into(),
-            prices: prices.into(),
+            invoice,
             start_parameter: None,
-            provider_data: None,
-            photo: None,
-            need_name: None,
-            need_phone_number: None,
-            need_email: None,
-            need_shipping_address: None,
-            send_phone_number_to_provider: None,
-            send_email_to_provider: None,
-            is_flexible: None,
             disable_notification: None,
             reply_to_message_id: None,
             allow_sending_without_reply: false,
             reply_markup: None,
-            tip: None,
         }
-    }
-
-    /// Configures data for your payment provider.
-    /// Reflects the `provider_data` parameter.
-    pub fn provider_data(mut self, provider_data: impl Into<String>) -> Self {
-        self.provider_data = Some(provider_data.into());
-        self
-    }
-
-    /// Configures a photo for the invoice.
-    /// Reflects the `photo_url`, `photo_width` and `photo_height` parameters.
-    #[allow(clippy::missing_const_for_fn)]
-    pub fn photo(mut self, photo: Photo) -> Self {
-        self.photo = Some(photo);
-        self
-    }
-
-    /// Configures whether the user must specify their name.
-    /// Reflects the `need_name` parameters.
-    pub const fn is_name_needed(mut self, is_needed: bool) -> Self {
-        self.need_name = Some(is_needed);
-        self
-    }
-
-    /// Configures whether the user must specify their phone number.
-    /// Reflects the `need_phone_number` parameter.
-    pub const fn is_phone_number_needed(mut self, is_needed: bool) -> Self {
-        self.need_phone_number = Some(is_needed);
-        self
-    }
-
-    /// Configures whether the user must specify their email.
-    /// Reflects the `need_email` parameter.
-    pub const fn is_email_needed(mut self, is_needed: bool) -> Self {
-        self.need_email = Some(is_needed);
-        self
-    }
-
-    /// Configures whether the user must specify their shipping address.
-    /// Reflects the `need_shipping_address` parameter.
-    pub const fn is_shipping_address_needed(mut self, is_needed: bool) -> Self {
-        self.need_shipping_address = Some(is_needed);
-        self
-    }
-
-    /// Configures whether the user's phone must be sent to your payment
-    /// provider. Reflects the `send_phone_number_to_provider` parameter.
-    pub const fn should_send_phone_number_to_provider(
-        mut self,
-        must_send: bool,
-    ) -> Self {
-        self.send_phone_number_to_provider = Some(must_send);
-        self
-    }
-
-    /// Configures whether the user's email must be sent to your payment
-    /// provider. Reflects the `send_email_to_provider` parameter.
-    pub const fn should_send_email_to_provider(
-        mut self,
-        must_send: bool,
-    ) -> Self {
-        self.send_email_to_provider = Some(must_send);
-        self
-    }
-
-    /// Configures whether the final price depends on the shipping method.
-    /// Reflects the `is_flexible` parameter.
-    pub const fn is_flexible(mut self, is_flexible: bool) -> Self {
-        self.is_flexible = Some(is_flexible);
-        self
     }
 
     /// Configures whether the message is sent silently.
@@ -206,14 +91,6 @@ impl<'a> SendInvoice<'a> {
         start_parameter: impl Into<String>,
     ) -> Self {
         self.start_parameter = Some(start_parameter.into());
-        self
-    }
-
-    /// Configures tip for the invoice.
-    /// Reflects the `tip` parameter.
-    #[allow(clippy::missing_const_for_fn)]
-    pub fn tip(mut self, tip: tip::Tip) -> Self {
-        self.tip = Some(tip);
         self
     }
 }
