+++ src/types/input_message_content.rs
@@ -4,13 +4,17 @@ use is_macro::Is;
 use serde::Serialize;
 
 mod contact;
+mod invoice;
 mod location;
 mod text;
 mod venue;
 
-pub use {contact::Contact, location::Location, text::Text, venue::Venue};
+pub use {
+    contact::Contact, invoice::Invoice, location::Location, text::Text,
+    venue::Venue,
+};
 
-/// Represents [`InputMessageContext`][docs].
+/// Represents [`InputMessageContent`][docs].
 ///
 /// [docs]: https://core.telegram.org/bots/api#inputmessagecontent
 #[derive(Debug, PartialEq, Clone, Serialize, Is)]
@@ -25,6 +29,8 @@ pub enum InputMessageContent {
     Venue(Venue),
     /// A contact.
     Contact(Contact),
+    /// An invoice.
+    Invoice(Invoice),
 }
 
 impl From<Text> for InputMessageContent {
@@ -52,5 +58,12 @@ impl From<Contact> for InputMessageContent {
     #[must_use]
     fn from(contact: Contact) -> Self {
         Self::Contact(contact)
+    }
+}
+
+impl From<Invoice> for InputMessageContent {
+    #[must_use]
+    fn from(invoice: Invoice) -> Self {
+        Self::Invoice(invoice)
     }
 }
