+++ src/types/parameters.rs
@@ -3,16 +3,19 @@
 mod allowed_updates;
 mod callback_action;
 mod chat_id;
+mod invoice;
 mod photo;
 pub mod poll;
 mod text;
-pub mod tip;
+mod tip;
 
 pub(crate) use text::ParseMode;
 pub use {
     allowed_updates::AllowedUpdates,
     callback_action::CallbackAction,
     chat_id::{ChatId, ImplicitChatId},
+    invoice::Invoice,
     photo::Photo,
     text::Text,
+    tip::Tip,
 };
